The R script used in the winter project conduct by Hungwei Lin
The R script used in the winter project conduct by Hungwei Lin, the project name is "The evolution of flight adaptations in non-passerine birds: disentangling the roles of climate and elevation". If you have any questions about the content of the code, please send an email to "homewaylin@qq.com".